Add tests for update actions. #925

+++ sphere-models/src/main/java/io/sphere/sdk/reviews/commands/updateactions/SetRating.java
@@ -5,21 +5,26 @@ import io.sphere.sdk.reviews.Review;
 
 import javax.annotation.Nullable;
 
+/**
+ * Sets/unsets the rating of a review.
+ *
+ * {@include.example io.sphere.sdk.reviews.commands.ReviewUpdateCommandTest#setRating()}
+ */
 public class SetRating extends UpdateActionImpl<Review> {
     @Nullable
-    private final Double rating;
+    private final Integer rating;
 
-    private SetRating(@Nullable final Double rating) {
+    private SetRating(@Nullable final Integer rating) {
         super("setRating");
         this.rating = rating;
     }
 
-    public static SetRating of(@Nullable final Double rating) {
+    public static SetRating of(@Nullable final Integer rating) {
         return new SetRating(rating);
     }
 
     @Nullable
-    public Double getRating() {
+    public Integer getRating() {
         return rating;
     }
 }
